streamlit/requirements.txt
streamlit==1.39.0
streamlit-lottie
openpyxl
pandas
numpy
matplotlib
seaborn
python-dotenv